+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- fix instance class validation bug ([#9](https://github.com/seaofvoices/generator-luau/pull/9))
+
 ## 0.1.3
 
 - fix tag configuration effect cleanup to happen only once ([#7](https://github.com/seaofvoices/generator-luau/pull/7))